requirements.txt
requests
pyserial==3.5
pyTelegramBotAPI==3.7.6